refactor: infrastructure 패키지에 세부 구현만 두는 것으로 변경
