+++ src/main/java/shopping/controller/OrderController.java
@@ -5,7 +5,7 @@ import org.springframework.web.bind.annotation.*;
 import shopping.dto.response.OrderCreateResponse;
 import shopping.dto.response.OrderResponse;
 import shopping.dto.response.OrderResponses;
-import shopping.infrastructure.ExchangeRateApi;
+import shopping.service.ExchangeRateService;
 import shopping.service.OrderService;
 
 import static java.time.LocalDateTime.now;
@@ -14,11 +14,11 @@ import static java.time.LocalDateTime.now;
 public class OrderController {
 
     private final OrderService orderService;
-    private final ExchangeRateApi exchangeRateApi;
+    private final ExchangeRateService exchangeRateService;
 
-    public OrderController(final OrderService orderService, final ExchangeRateApi exchangeRateApi) {
+    public OrderController(final OrderService orderService, final ExchangeRateService exchangeRateService) {
         this.orderService = orderService;
-        this.exchangeRateApi = exchangeRateApi;
+        this.exchangeRateService = exchangeRateService;
     }
 
     @GetMapping("/orders/{orderId}")
@@ -37,7 +37,7 @@ public class OrderController {
 
     @PostMapping("/orders")
     public ResponseEntity<OrderCreateResponse> createOrder(@RequestAttribute final Long loginMemberId) {
-        final double exchangeRate = exchangeRateApi.getExchangeRateEveryMinute(now());
+        final double exchangeRate = exchangeRateService.getExchangeRateEveryMinute(now());
         final OrderCreateResponse orderCreateResponse = orderService.createOrder(loginMemberId, exchangeRate);
 
         return ResponseEntity.ok(orderCreateResponse);
