+++ src/test/java/shopping/config/TestConfig.java
@@ -3,14 +3,14 @@ package shopping.config;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.context.annotation.Profile;
-import shopping.infrastructure.ExchangeRateApi;
+import shopping.service.ExchangeRateService;
 
 @Configuration
 public class TestConfig {
 
     @Bean
     @Profile("test")
-    public ExchangeRateApi exchangeRateApi() {
+    public ExchangeRateService exchangeRateApi() {
         return (time) -> 1300.0;
     }
 }
